Changed test setup
- Changed IListenableTest from an interface to an abstract class, since it
  was causing problems.
- Let TestGameThread extend TestListenable.
- Created TestRenderer2 that extends TestListenable.
- Added some missing documentation

<testing,documentation
